refactor(secondary-screen): promote achievement comments page to a widget
Extract _buildAchievementCommentsPage and its comment-card / load-more /
message / date-format sub-builders out of _SecondaryScreenState into a new
const-ctor AchievementCommentsPage under secondary_screen/widgets/. The host
passes the selected achievement, its cached comment state snapshot, the l10n
context, and the load callback down — the child re-reads no state of its own.

The private _AchievementCommentsState loading-state holder is promoted to a
public AchievementCommentsState in the same file (the host's _commentsCache and
_loadAchievementComments construct/reference it); bodies moved verbatim aside
from the intentional _l10nContext -> l10nContext and _loadAchievementComments
-> onLoadComments substitutions. Completes B's panel extractions.
